Factor out the bulk of the react components for scene explorer (improves fast refresh)
